+++ app/admin/__init__.py
@@ -1,12 +1,11 @@
 from .user_admin import ClientAdmin, EmployeeAdmin
 from .offer_admin import OfferLinkAdmin
 from .order_admin import OrderAdmin, OrderDetailAdmin
-from .service_admin import ServiceNameAdmin, ServiceCategoryAdmin
+from .service_admin import ServiceNameAdmin
 
 admin_classes = (
     EmployeeAdmin,
     ClientAdmin,
-    ServiceCategoryAdmin,
     ServiceNameAdmin,
     OfferLinkAdmin,
     OrderAdmin,
