+++ app/admin/service_admin.py
@@ -1,18 +1,6 @@
 from sqladmin import ModelView
 
-from models import CategoryModel, ServiceNameModel
-
-
-class ServiceCategoryAdmin(ModelView, model=CategoryModel):
-    column_list = [
-        CategoryModel.id,
-        CategoryModel.created_at,
-        CategoryModel.changed_at,
-        CategoryModel.name,
-        CategoryModel.is_active,
-    ]
-    column_searchable_list = [CategoryModel.name]
-    column_details_exclude_list = [CategoryModel.services]
+from models import ServiceNameModel
 
 
 class ServiceNameAdmin(ModelView, model=ServiceNameModel):
