+++ app/includer/admins.py
@@ -1,0 +1,6 @@
+from category.admin_category import ServiceCategoryAdmin
+
+
+admin_classes2 = (
+    ServiceCategoryAdmin,
+)
