+++ app/includer/routers.py
@@ -1,0 +1,5 @@
+from category.category_routers import router_category
+
+routers_all2 = [
+    router_category,
+]
