+++ app/main.py
@@ -3,16 +3,22 @@ from sqladmin import Admin
 
 from admin import admin_classes
 from config import i_config
+from includer.admins import admin_classes2
+from includer.routers import routers_all2
 from routers import routers_all
 from models.database import engine
 
 app = FastAPI(debug=i_config.DEBUG)
 
 admin = Admin(app, engine)
+for admin_class in admin_classes2:
+    admin.add_view(admin_class)
 for admin_class in admin_classes:
     admin.add_view(admin_class)
 
 for route in routers_all:
+    app.include_router(route)  # , prefix='/api/auth', tags=['auth']
+for route in routers_all2:
     app.include_router(route)  # , prefix='/api/auth', tags=['auth']
