+++ app/models/service_model.py
@@ -2,25 +2,13 @@ from typing import TYPE_CHECKING, List, Union
 
 from pydantic import constr, condecimal
 from sqlmodel import Field, Relationship
-from sqlalchemy import VARCHAR, Column, UniqueConstraint
+from sqlalchemy import UniqueConstraint
 
+from category.model_category import CategoryModel
 from models.base_models import BaseSQLModel, DateCreatedChangedBase
 
 if TYPE_CHECKING:
     from models import OrderModel, OfferLinkModel
-
-
-class CategoryInSchema(BaseSQLModel):
-    name: constr(min_length=2, max_length=50) = Field(sa_column=Column('name', VARCHAR, unique=True))
-    is_active: Union[bool, None] = True
-    detail: constr(max_length=50) = ''
-
-
-class CategoryModel(DateCreatedChangedBase, CategoryInSchema, table=True):
-    __tablename__ = 'category'
-    id: int = Field(default=None, primary_key=True)
-
-    services: List['ServiceNameModel'] = Relationship(back_populates='category')
 
 
 class ServiceNameInSchema(BaseSQLModel):
