+++ app/routers/__init__.py
@@ -4,7 +4,6 @@ from routers.offer_routers import router_offer
 from routers.order_routers import router_order
 from routers.client_routers import router_client
 from routers.service_routers import router_service
-from routers.category_routers import router_category
 from routers.employee_routers import router_employee
 from routers.stub_init_routers import router_init_stub
 
@@ -12,7 +11,6 @@ routers_all = [
     router_index,
     router_employee,
     router_client,
-    router_category,
     router_service,
     router_offer,
     router_order,
